requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
langchain>=0.1.0
langchain-openai>=0.0.5
python-dotenv>=1.0.0
plotly>=5.15.0
openpyxl>=3.1.0
zipfile36>=0.1.3
chardet>=5.0.0
openai>=1.0.0
langgraph>=0.1.0 